VMS-63086: Fix deprecated QHashCombiner usage

+++ vms/client/nx_vms_client_desktop/src/nx/vms/client/desktop/shaders/media_output_shader_program.h
@@ -88,6 +88,6 @@ private:
     const Key m_key;
 };
 
-size_t qHash(const MediaOutputShaderProgram::Key& key, uint seed = 0);
+size_t qHash(const MediaOutputShaderProgram::Key& key, size_t seed = 0);
 
 } // namespace nx::vms::client::desktop
